Change behavior of the second long menu item
Normally long running events alternate between beginning and ending,
and the point of the second menu item is to break of the cycle when need,
e.g. when missing one or when the event is created while it is running.

To break the cycle, the second menu item can either silently toggle
between running and non-running, or record normally the opposite of current
event. Previously, the code implemented the former while being labelled as
the latter, which is probably not a good thing.

The second idea seems slightly easier for the operator (who is already
being punished by having to scroll down to the bottom of the event list),
and much easier to come up with a clear label. So the code has been
aligned to the label.